% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{get_doc}
\alias{get_doc}
\title{Formats a table of variable names}
\usage{
get_doc(o)
}
\arguments{
\item{o}{is the ipysd S3 object}
}
\value{
tibble
}
\description{
\code{get_doc()} Get mode variable names
}
\examples{
\dontrun{
target <- system.file("models/vensim", "Population.mdl", package = "pysd2r")
py <- pysd_connect()
py <- read_vensim(py, target)
mdoc <- get_doc(py)
}
}
