% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{get_final_time}
\alias{get_final_time}
\title{Gets the final time from the model}
\usage{
get_final_time(o)
}
\arguments{
\item{o}{is the ipysd S3 object}
}
\value{
The finaltime
}
\description{
\code{get_timestep} uses pysd to fetch the time step from the model
}
\details{
As it's a generic function, this call is dispatched to set_component.isdpy
}
