% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{get_python_info}
\alias{get_python_info}
\title{Creates an object to facilitate interaction with pysd}
\usage{
get_python_info()
}
\value{
python information
}
\description{
\code{get_python_version} returns information on what version of python
is being used with reticulate
}
