% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{pysd_connect}
\alias{pysd_connect}
\title{Creates an object to facilitate interaction with pysd}
\usage{
pysd_connect()
}
\value{
An S3 object of class ipysd
}
\description{
\code{pysd_connect} returns a ipysd object to the calling program.
This object will contain a link variable to pysd and will subsequently
store a reference to the simulation model in pysd.
}
\details{
\href{https://pysd.readthedocs.io/en/master/}{Link to pysd}

The result is used as a parameter for read_vensim() & read_xmile() functions
}
\examples{
\dontrun{
py pysd_connect()
}
}
