% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{reload_model}
\alias{reload_model}
\title{Reloads the model from original mdl file}
\usage{
reload_model(o)
}
\arguments{
\item{o}{is the ipysd S3 object}
}
\value{
ipysd object
}
\description{
\code{reload_model()} Reloads the model
}
