% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{run_model}
\alias{run_model}
\title{Runs a simulation model}
\usage{
run_model(o)
}
\arguments{
\item{o}{is the ipysd S3 object}
}
\value{
tibble containing the simulation results
}
\description{
\code{run_model()} calls \code{run} in pysd and returns all
the simulation output in tidy data format (tibble)
}
\details{
As it's a generic function, this call is dispatched to run_model.isdpy
}
\examples{
\dontrun{
target <- system.file("models/vensim", "Population.mdl", package = "pysd2r")
py <- pysd_connect()
py <- read_vensim(py, target)
results <- run_model(py)
}
}
