% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{set_time_values}
\alias{set_time_values}
\title{Sets the initial time, final time, and timestep}
\usage{
set_time_values(o, init, final, DT)
}
\arguments{
\item{o}{is the ipysd S3 object}

\item{init}{is the initial time}

\item{final}{is the final time}

\item{DT}{is the time step}
}
\description{
\code{set_time_valuesl()} sets the simulation times and DT
}
\examples{
\dontrun{
target <- system.file("models/vensim", "Population.mdl", package = "pysd2r")
py <- pysd_connect()
py <- read_vensim(py, target)
set_time_values(py,0,10,0.5)
}
}
