% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{get_python_info}
\alias{get_python_info}
\title{Gets the current python configuration for reticulate}
\usage{
get_python_info()
}
\value{
python information
}
\description{
\code{get_python_info} returns information on what version of python
is being used with reticulate
}
\examples{
\dontrun{
get_python_info()
}
}
