% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipysd.R
\name{reload_model}
\alias{reload_model}
\title{Reloads the model from original mdl file}
\usage{
reload_model(o)
}
\arguments{
\item{o}{is the ipysd S3 object}
}
\value{
ipysd object
}
\description{
\code{reload_model()} Reloads the model
}
\examples{
\dontrun{
target <- system.file("models/vensim", "Population.mdl", package = "pysd2r")
py <- pysd_connect()
py <- read_vensim(py, target)
set_time_values(py,0,10,0.5)
py<-reload_model(py)
}
}
